Various updates to stream the extension.
