+++ pyblish_deadline/examples/extract_deadline_draft.py
@@ -17,8 +17,8 @@ class ExtractDeadlineDraft(pyblish.api.Extractor):
 
         # getting job data
         job_data = {}
-        if instance.has_data('deadlineJobData'):
-            job_data = instance.data('deadlineJobData').copy()
+        if instance.has_data('deadlineData'):
+            job_data = instance.data('deadlineData')['job'].copy()
 
         # setting extra info key values
         extra_info_key_value = {}
@@ -44,4 +44,6 @@ class ExtractDeadlineDraft(pyblish.api.Extractor):
 
         job_data['ExtraInfoKeyValue'] = extra_info_key_value
 
-        instance.set_data('deadlineJobData', value=job_data)
+        data = instance.data('deadlineData')
+        data['job'] = job_data
+        instance.set_data('deadlineData', value=data)
